+++ frontend/astro.config.mjs
@@ -1,0 +1,11 @@
+import { defineConfig } from 'astro/config';
+import tailwind from "@astrojs/tailwind";
+
+// import cloudflare from "@astrojs/cloudflare";
+
+// https://astro.build/config
+export default defineConfig({
+  integrations: [tailwind()],
+  output: 'static',
+  // adapter: cloudflare({ mode: "directory" }),
+});
