--- frontend/index.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>にじ通</title>
-  <link rel="stylesheet" href="style.css">
-  <link rel="manifest" href="manifest.webmanifest">
-  <meta name="description" content="にじさんじの歌みた動画を通知するサービスです">
-</head>
-
-<script src='main.js' type='module'></script>
-
-<body>
-  <h1>にじ通</h1>
-  <details>
-    <summary>メニュー</summary>
-    <p><a href="index.html">ホーム</a></p>
-    <p><a href="how-to.html">使い方</a></p>
-    <p><a href="terms.html">利用規約・プライバシーポリシー</a></p>
-    <p><a href="form.html">問い合わせ</a></p>
-  </details>
-  <hr align="LEFT" width="100%">
-
-  <div id="overlay">
-    <div id="text">
-      <span id="loader"></span>
-    </div>
-  </div>
-
-  <dialog id="favDialog">
-    <p>
-      にじ通へようこそ！<br />
-      当サイトをお使いの前に「<a href="terms.html">利用規約・プライバシーポリシー</a>」をご確認ください。
-    </p>
-    <input id='ok' type="button" value="確認しました">
-  </dialog>
-
-  <dialog id="WarningDialog">
-    <p>
-      iPhone, iPad, iPodをお使いの場合、「共有」→「ホーム画面に追加」して、追加されたアイコンから起動してください
-    </p>
-  </dialog>
-
-  <div id="subscription-conf">
-    <h3>通知設定</h3>
-    <div style="width:100%; display:flex; justify-content:flex-start;">
-      <input id='song' type="checkbox" value="">
-      <p style="margin-left: 1em;">歌みた動画が公開される5分前に通知を受け取る</p>
-    </div>
-    <div style="width:100%; display:flex; justify-content:flex-start;">
-      <input id='keyword' type="checkbox" value="">
-      <p style="margin-left: 1em;">指定したキーワードを含む動画がアップロードされた際に通知を受け取る</p>
-    </div>
-    <input id='keyword-text' type="search" value="" style="margin-left: 3em;">
-    <br />
-    <br />
-    <input id='submit' type="button" value="保存">
-    <input id='delete' type="button" value="通知解除">
-  </div>
-</body>
-
-</html>
