--- frontend/public/form.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>にじ通</title>
-  <link rel="stylesheet" href="style.css">
-  <link rel="manifest" href="manifest.webmanifest">
-</head>
-
-<body>
-  <h1>にじ通</h1>
-  <details>
-    <summary>メニュー</summary>
-    <p><a href="index.html">ホーム</a></p>
-    <p><a href="how-to.html">使い方</a></p>
-    <p><a href="terms.html">利用規約・プライバシーポリシー</a></p>
-    <p><a href="form.html">問い合わせ</a></p>
-  </details>
-  <hr align="LEFT" width="100%">
-
-  <h3>問い合わせ</h3>
-  <p>問い合わせはGoogleフォーム、X(Twitter)までお願いします_(._.)_</p>
-  <ul>
-    <li><a href="https://forms.gle/1hGxTx8v1sN8FnUC9" target="_blank" rel="noopener noreferrer">Google フォーム</a></li>
-    <li><a href="https://twitter.com/zxZ5oETkh75KKlj" target="_blank" rel="noopener noreferrer">X(Twitter)</a></li>
-  </ul>
-  
-</body>
-
-</html>
