--- frontend/public/how-to.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>にじ通</title>
-  <link rel="stylesheet" href="style.css">
-  <link rel="manifest" href="manifest.webmanifest">
-</head>
-
-<body>
-  <h1>にじ通</h1>
-  <details>
-    <summary>メニュー</summary>
-    <p><a href="index.html">ホーム</a></p>
-    <p><a href="how-to.html">使い方</a></p>
-    <p><a href="terms.html">利用規約・プライバシーポリシー</a></p>
-    <p><a href="form.html">問い合わせ</a></p>
-  </details>
-  <hr align="LEFT" width="100%">
-
-  <h3>使い方</h3>
-  <p>
-    通知を受け取りたい項目にチェックし、「保存」ボタンを押すことで通知を受け取ることができるようになります。<br />
-    このサービスはPush通知が用いており、通知はお使いのデバイスのロック画面やステータスバーなどに表示されます。
-  </p>
-  <h4>歌みた動画が公開される5分前に通知を受け取る</h4>
-  <li>プレミア公開される歌ってみた動画が公開される5分前に通知されます。</li>
-  <h4>指定したキーワードを含む動画がアップロードされた際に通知を受け取る</h4>
-  <li>対象の項目の下にある入力欄にキーワードを入力することで、指定したキーワードを含む動画タイトルがアップロードされたときに通知されます。</li>
-  
-</body>
-
-</html>
