--- frontend/public/terms.html
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>にじ通</title>
-  <link rel="stylesheet" href="style.css">
-  <link rel="manifest" href="manifest.webmanifest">
-</head>
-
-<body>
-  <h1>にじ通</h1>
-  <details>
-    <summary>メニュー</summary>
-    <p><a href="index.html">ホーム</a></p>
-    <p><a href="how-to.html">使い方</a></p>
-    <p><a href="terms.html">利用規約・プライバシーポリシー</a></p>
-    <p><a href="form.html">問い合わせ</a></p>
-  </details>
-  <hr align="LEFT" width="100%">
-
-  <h2>利用規約</h2>
-  <p>
-    にじ通（以下、「本サービス」と言います）に関する利用規約（以下『本規約』）を以下の通り定めます。<br />
-    本サービスご利用にあたっては、本規約とプライバシーポリシーの内容をお読みいただき承諾したものとします。<br />
-  </p>
-
-  <h4>権利帰属</h4>
-  <p>
-    本サービスは、ANYCOLOR株式会社様の公式コンテンツではありません。<br />
-    本サービスから通知されるチャンネル情報、配信情報、画像の著作権や肖像権等は、当該権利所有者様に帰属しています。
-  </p>
-
-  <h4>免責事項</h4>
-  <p>
-    本サービスは、配信情報、チャンネル情報などすべての情報が、正確かつ最新であることを保証しておりません。<br />
-  </p>
-
-  <h4>禁止事項</h4>
-  <li>本サービスの運営を妨害するおそれのある行為</li>
-  <li>不正アクセスをし，またはこれを試みる行為</li>
-  <li>その他，本サービスの管理者が不適切と判断する行為</li>
-  
-  <h4>本サービスの提供の停止等</h4>
-  <p>以下のいずれかの事由があると判断した場合，本サービスの全部または一部の提供を停止または中断することができるものとします。</p>
-  <li>本サービスにかかるコンピュータシステムの保守点検または更新を行う場合</li>
-  <li>地震，落雷，火災，停電または天災などの不可抗力により，本サービスの提供が困難となった場合</li>
-  <li>コンピュータまたは通信回線等が事故により停止した場合</li>
-  <li>ANYCOLOR株式会社様、ANYCOLOR株式会社所属のライバー様からサービスに関するお申し出を受けた場合</li>
-  <li>その他，当社が本サービスの提供が困難と判断した場合</li>
-  <p>本サービスの提供の停止または中断により，ユーザーまたは第三者が被ったいかなる不利益または損害についても，一切の責任を負わないものとします。</p>
-
-  <h4>サービス内容の変更等</h4>
-  <p>本サービスは、ユーザーへの事前の告知をもって、本サービスの内容を変更、追加または廃止することがあり、ユーザーはこれを承諾するものとします。</p>
-
-  <h4>利用規約の変更</h4>
-  <p>必要に応じて、この利用規約の内容を変更します。この場合、変更後の利用規約の施行時期と内容を適切な方法により周知または通知します。</p>
-
-  <h2>プライバシーポリシー</h2>
-  <p>
-    本サービスは、機能を提供・向上させるために最低限の個人情報を取り扱います。
-  </p>
-  <h4>本サービスが取り扱う情報</h4>
-  <li>FCMトークン</li>
-  <li>OS情報</li>
-  <li>ブラウザ情報</li>
-  <li>ユーザーが入力したキーワード</li>
-  <li>IPアドレス</li>
-  <li>問い合わせ内容</li>
-
-  <h4>利用目的</h4>
-  <li>通知に必要な情報、OS情報、ブラウザ情報、ユーザーが入力したキーワードは、本サービスを提供する目的のみに使用します。</li>
-  <li>
-    また、FCMトークンに関しては、上記の目的以外に、
-    本サービスの変更、提供中止、終了、
-    本サービスの規約の変更等を通知
-    の目的にも使用します。
-  </li>
-  <li>IPアドレスは、本サービスの規約に違反する行為をしたユーザーのみのIPアドレスを収集し、違反行為に対応する目的のみに使用します。</li>
-  <li>問い合わせ内容は、本サービスの維持改善に役立てる目的のみに使用します。</li>
-
-  <h4>FCMトークンについて</h4>
-  <p>
-    本サービスは、FCMトークンを収集します。<br />
-    本サービスの通知を許可することで、本サービスのサーバーにFCMトークンが送信されます。<br />
-    FCMトークンには、ユーザーの氏名やメールアドレスなど、個人を特定できる情報は含まれておらず、FCMトークンを使ってユーザーのデバイスに通知を送ることしかできません。<br />
-    また、FCMトークンはユーザー側で通知のブロック、通知の解除をすることでいつでも無効化することができます。<br />
-  </p>
-
-  <h4>JavaScriptの利用について</h4>
-  <p>
-    本サービスでは、通知を受け取るようにするため、JavaScriptを使用しております。<br />
-    JavaScriptを無効にすると、本サービスをご利用できませんので、JavaScriptを有効にしていただけますようお願いいたします。<br />
-  </p>
-
-  <h4>問い合わせについて</h4>
-  <p>
-    Googleフォームから問い合わせの際、問い合わせ内容が本サービスのサーバーに送信されます。
-    問い合わせの際、Googleアカウントが必要ですが、ユーザーのGoogleアカウント情報は収集いたしません。
-  </p>
-
-  <h4>個人情報の第三者提供</h4>
-  <p>ユーザーからお預かりした個人情報を、個人情報保護法その他の法令に基づき開示が認められる場合を除き、ご本人様の同意を得ずに第三者に提供することはありません。</p>
-
-  <h4>本サービスのコンテンツ</h4>
-  <p>
-    本サービスは、YouTube APIサービスによって動画情報の取得・通知をおこなっています。<br />
-    また、YouTube APIサービスの利用は、YouTubeの「<a href="https://www.youtube.com/t/terms" target="_blank" rel="noopener noreferrer">利用規約</a>」
-    及びGoogleの「<a href="https://policies.google.com/privacy" target="_blank" rel="noopener noreferrer">プライバシーポリシー</a>」に則り利用しております。<br />
-    ユーザーは本サービスを利用することで「<a href="https://www.youtube.com/t/terms" target="_blank" rel="noopener noreferrer">YouTube利用規約</a>」及び「<a href="https://www.youtube.com/t/terms" target="_blank" rel="noopener noreferrer">Googleプライバシーポリシー</a>」に同意したものとみなされます。
-  </p>
-
-  <h4>プライバシーポリシーの変更</h4>
-  <p>必要に応じて、このプライバシーポリシーの内容を変更します。この場合、変更後のプライバシーポリシーの施行時期と内容を適切な方法により周知または通知します。</p>
-
-  <br />
-  <p>2024年05月12日 制定</p>
-</body>
-
-</html>
