+++ frontend/src/components/SubmitButton.astro
@@ -1,0 +1,157 @@
+<button
+    id="submit"
+    type="button"
+    class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center me-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800 inline-flex items-center"
+>
+    <div id="submit-spinner" class="hidden">
+        <svg
+            aria-hidden="true"
+            role="status"
+            class="inline w-4 h-4 me-3 text-white animate-spin"
+            viewBox="0 0 100 101"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+        >
+            <path
+                d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
+                fill="#E5E7EB"></path>
+            <path
+                d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
+                fill="currentColor"></path>
+        </svg>
+    </div>
+    <div id="submit-text">通知登録</div>
+</button>
+
+<script>
+    import { getMessaging, getToken } from "firebase/messaging";
+    import { vapidKey } from "../scripts/config";
+
+    const messaging = getMessaging();
+    const submitEle = document.getElementById("submit") as HTMLElement;
+    const submitSpinnerEle = document.getElementById("submit-spinner") as HTMLElement;
+    const submitTextEle = document.getElementById("submit-text") as HTMLElement;
+    const toastEle = document.getElementById("toast-success") as HTMLElement;
+
+    submitEle.addEventListener("click", async () => {
+        console.log("POST");
+
+        indexedDBSet()
+
+        const songChecked = window.localStorage.getItem("song-checked") == "true" ? 1 : 0;
+        const keywordChecked = window.localStorage.getItem("keyword-checked") == "true" ? 1 : 0;
+
+        submitSpinnerEle.style.display = "block";
+        submitTextEle.innerText = "通知登録中...";
+
+        console.log("generate token...");
+        const currentToken = await getFCMToken();
+        if (currentToken == "") {
+            submitSpinnerEle.style.display = "none";
+            submitTextEle.innerText = "通知登録";
+            return;
+        }
+
+        const ok = await saveFCMToken(currentToken, {
+            song: songChecked,
+            word: keywordChecked,
+        });
+        if (!ok) {
+            window.alert("トークンの登録に失敗しました");
+            submitSpinnerEle.style.display = "none";
+            submitTextEle.innerText = "通知登録";
+            return;
+        }
+
+        window.localStorage.setItem("fcm-token", currentToken);
+
+        toastEle.style.visibility = "visible";
+        submitSpinnerEle.style.display = "none";
+        submitTextEle.innerText = "通知登録";
+    });
+
+    const getFCMToken = async () => {
+        try {
+            const currentToken = await getToken(messaging, { vapidKey });
+            console.log("generated token:", currentToken);
+            return currentToken;
+        } catch (error) {
+            // 通知権限がブロックされている場合
+            if (Notification.permission === "denied") {
+                window.alert("通知がブロックされています。");
+            }
+
+            // 通知権限がブロックされていないが、ユーザーの許可を得れていない場合
+            if (Notification.permission === "default") {
+                window.alert("通知を許可してください。");
+            }
+
+            // 通知権限がブロックされていないが、ユーザーの許可を得れていない場合
+            if (Notification.permission === "granted") {
+                window.alert(
+                    "エラーが発生しました。再度通知登録をしてください。",
+                );
+            }
+        }
+        return "";
+    };
+
+    const saveFCMToken = async (token: string, data = {}) => {
+        const url = "/api/subscription";
+        const response = await fetch(url, {
+            method: "POST",
+            cache: "no-cache",
+            headers: {
+                "Content-Type": "application/json",
+                Token: JSON.stringify(token),
+            },
+            body: JSON.stringify(data), // 本体のデータ型は "Content-Type" ヘッダーと一致させる必要があります
+        });
+
+        console.log(response.status);
+        console.log(response.ok);
+        return response.ok;
+    };
+
+    const indexedDBSet = () => {
+        const word = window.localStorage.getItem("word");
+        if (word == null || word == "") return;
+
+        const openRequest = window.indexedDB.open("niji-tuu", 1);
+
+        openRequest.onupgradeneeded = function () {
+            console.log("onupgradeneeded")
+            const db = openRequest.result;
+            if (!db.objectStoreNames.contains("keyword")) {
+                console.log("objectStoreNames")
+                db.createObjectStore("keyword", { keyPath: "id" });
+            }
+            // valueSet(db, word)
+        };
+
+        openRequest.onerror = function () {
+            console.error("indexedDB Error");
+            window.alert("indexedDB Error");
+        };
+
+        openRequest.onsuccess = function () {
+            console.log("onsuccess")
+            const db = openRequest.result;
+            valueSet(db, word)
+        };
+    };
+
+    const valueSet = (db: IDBDatabase, value: string) => {
+        const transaction = db.transaction("keyword", "readwrite");
+        const store = transaction.objectStore("keyword")
+        const req = store.put({"id": "1", "text": value})
+        req.onsuccess = (event) => {
+            // event.target.result === customer.ssn;
+            console.log("onsuccess:", event)
+        };
+        req.onerror = (event) => {
+            console.log("onerror:", event)
+        }
+
+    }
+</script>
