+++ frontend/src/components/keyWord.astro
@@ -1,0 +1,15 @@
+<input
+    type="text"
+    id="keyword-text"
+    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full ps-5 p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 mt-2"
+    placeholder="キーワード..."
+/>
+
+<script>
+    const ele = document.getElementById("keyword-text") as HTMLElement
+
+    ele.addEventListener("change", (event) => {
+        window.localStorage.setItem("word", event.target?.value)
+    })
+    
+</script>
