+++ frontend/src/components/tearmsDialog.astro
@@ -1,0 +1,25 @@
+<dialog id="favDialog" class="relative bg-white rounded-lg shadow dark:bg-gray-700 absolute -translate-y-1/2 top-2/4">
+    <div class="p-4 md:p-5 space-y-4">
+        <p class="text-base leading-relaxed text-gray-500 dark:text-gray-400">
+            にじ通へようこそ！<br />
+            当サイトをお使いの前に「<a href="/terms">利用規約・プライバシーポリシー</a>」をご確認ください。
+        </p>
+    </div>
+    <!-- Modal footer -->
+    <div class="flex items-center p-4 md:p-5 border-t border-gray-200 rounded-b dark:border-gray-600">
+        <button id="ok" type="button" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">確認しました！</button>
+    </div>
+</dialog>
+
+<script>
+    const favDialog = document.getElementById("favDialog");
+
+    if (window.localStorage.getItem("readed-terms") !== "true") {
+        favDialog.showModal();
+    }
+
+    document.getElementById("ok").addEventListener("click", () => {
+        window.localStorage.setItem("readed-terms", "true")
+        favDialog.close();
+    });
+</script>
