+++ frontend/src/env.d.ts
@@ -1,0 +1,6 @@
+/// <reference types="astro/client" />
+type Runtime = import("@astrojs/cloudflare").Runtime<Env>;
+
+declare namespace App {
+	interface Locals extends Runtime {}
+}
