+++ frontend/src/layouts/Layout.astro
@@ -1,0 +1,84 @@
+<html>
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>にじ通</title>
+  <meta name="description" content="にじさんじの歌みた動画を通知するサービスです">
+  <link rel="manifest" href="manifest.webmanifest">
+</head>
+
+<body class="dark:bg-gray-800">
+	<header class="border-b-2 border-inherit">
+		<nav class="border-gray-200 bg-gray-50 dark:bg-gray-800 dark:border-gray-700">
+		  <div class="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
+			<a href="#" class="flex items-center space-x-3 rtl:space-x-reverse">
+			  <span class="self-center text-2xl font-semibold whitespace-nowrap dark:text-white">にじ通</span>
+			</a>
+			<button id="targetEl" data-collapse-toggle="triggerEl" type="button"
+			  class="inline-flex items-center justify-center p-2 w-10 h-10 text-sm text-gray-500 rounded-lg hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600"
+			  aria-controls="triggerEl" aria-expanded="false">
+			  <span class="sr-only">Open main menu</span>
+			  <svg class="w-5 h-5" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 17 14">
+				<path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+				  d="M1 1h15M1 7h15M1 13h15" />
+			  </svg>
+			</button>
+			<div class="hidden w-full" id="triggerEl">
+			  <ul class="flex flex-col font-medium mt-4 rounded-lg bg-gray-50 dark:bg-gray-800 dark:border-gray-700">
+				<li>
+				  <a href="/" class="block py-2 px-3 text-white bg-blue-700 rounded dark:bg-blue-600"
+					aria-current="page">ホーム</a>
+				</li>
+				<li>
+				  <a href="/how-to"
+					class="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white">使い方</a>
+				</li>
+				<li>
+				  <a href="/terms"
+					class="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 dark:text-gray-400 md:dark:hover:text-white dark:hover:bg-gray-700 dark:hover:text-white">利用規約・プライバシーポリシー</a>
+				</li>
+				<li>
+				  <a href="/form"
+					class="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white">問い合わせ</a>
+				</li>
+			  </ul>
+			</div>
+		  </div>
+		</nav>
+	</header>
+
+  	<main class="prose dark:prose-invert prose-h1:font-bold prose-h1:text-xl prose-a:text-blue-600 prose-p:text-justify prose-img:rounded-xl prose-headings:underline p-3">
+  		<slot/>
+	</main>
+</body>
+
+<script>
+    import { Collapse } from 'flowbite';
+
+    // set the target element that will be collapsed or expanded (eg. navbar menu)
+    const $targetEl = document.getElementById('targetEl');
+
+    // optionally set a trigger element (eg. a button, hamburger icon)
+    const $triggerEl = document.getElementById('triggerEl');
+
+    // optional options with default values and callback functions
+    const options = {
+        onCollapse: () => {
+            console.log('element has been collapsed');
+        },
+        onExpand: () => {
+            console.log('element has been expanded');
+        },
+        onToggle: () => {
+            console.log('element has been toggled');
+        },
+    };
+
+    const instanceOptions = {
+    id: 'targetEl',
+    override: true
+    };
+
+    const collapse = new Collapse($targetEl, $triggerEl, options, instanceOptions);
+</script>
+</html>
