+++ frontend/src/pages/form.md
@@ -1,0 +1,7 @@
+---
+layout: '../layouts/Layout.astro'
+---
+### 問い合わせ
+機能のリクエスト、バグ報告などは以下からお願いします
+- [Google フォーム](https://forms.gle/1hGxTx8v1sN8FnUC9)
+- [X(Twitter)](https://twitter.com/zxZ5oETkh75KKlj)
