+++ frontend/src/pages/how-to.md
@@ -1,0 +1,13 @@
+---
+layout: '../layouts/Layout.astro'
+---
+### 使い方
+通知を受け取りたい項目にチェックし、「保存」ボタンを押すことで通知を受け取ることができるようになります。  
+このサービスはPush通知を用いており、お使いのデバイスのロック画面やステータスバーなどに通知が表示されます。
+
+#### 通知設定
+**歌みた動画が公開される5分前に通知を受け取る:**
+プレミア公開される歌ってみた動画が公開される5分前に通知されます。 
+
+**指定したキーワードを含む動画がアップロードされた際に通知を受け取る:**
+対象の項目の下にある入力欄にキーワードを入力することで、指定したキーワードを含む動画タイトルがアップロードされたときに通知されます。
