+++ frontend/src/pages/terms.md
@@ -1,0 +1,78 @@
+---
+layout: '../layouts/Layout.astro'
+---
+## 利用規約
+にじ通（以下、「本サービス」と言います）に関する利用規約（以下『本規約』）を以下の通り定めます。
+本サービスご利用にあたっては、本規約とプライバシーポリシーの内容をお読みいただき承諾したものとします。
+
+#### 権利帰属
+本サービスは、ANYCOLOR株式会社様の公式コンテンツではありません。
+本サービスから通知されるチャンネル情報、配信情報、画像の著作権や肖像権等は、当該権利所有者様に帰属しています。
+    
+#### 免責事項
+本サービスは、配信情報、チャンネル情報などすべての情報が、正確かつ最新であることを保証しておりません。
+
+#### 禁止事項
+- 本サービスの運営を妨害するおそれのある行為
+- 不正アクセスをし，またはこれを試みる行為
+- その他，本サービスの管理者が不適切と判断する行為
+
+#### 本サービスの提供の停止等
+以下のいずれかの事由があると判断した場合，本サービスの全部または一部の提供を停止または中断することができるものとします。
+- 本サービスにかかるコンピュータシステムの保守点検または更新を行う場合
+- 地震，落雷，火災，停電または天災などの不可抗力により，本サービスの提供が困難となった場合
+- コンピュータまたは通信回線等が事故により停止した場合
+- ANYCOLOR株式会社様、ANYCOLOR株式会社所属のライバー様からサービスに関するお申し出を受けた場合
+- その他，当社が本サービスの提供が困難と判断した場合
+本サービスの提供の停止または中断により，ユーザーまたは第三者が被ったいかなる不利益または損害についても，一切の責任を負わないものとします。
+
+#### サービス内容の変更等
+本サービスは、ユーザーへの事前の告知をもって、本サービスの内容を変更、追加または廃止することがあり、ユーザーはこれを承諾するものとします。
+
+#### 利用規約の変更
+必要に応じて、この利用規約の内容を変更します。この場合、変更後の利用規約の施行時期と内容を適切な方法により周知または通知します。
+
+### プライバシーポリシー
+本サービスは、機能を提供・向上させるために最低限の個人情報を取り扱います。
+
+#### 本サービスが取り扱う情報
+- FCMトークン
+- OS情報
+- ブラウザ情報
+- ユーザーが入力したキーワード
+- IPアドレス
+- 問い合わせ内容
+
+#### 利用目的
+- 通知に必要な情報、OS情報、ブラウザ情報、ユーザーが入力したキーワードは、本サービスを提供する目的のみに使用します。
+- また、FCMトークンに関しては、上記の目的以外に、本サービスの変更、提供中止、終了、本サービスの規約の変更等を通知の目的にも使用します。
+- IPアドレスは、本サービスの規約に違反する行為をしたユーザーのみのIPアドレスを収集し、違反行為に対応する目的のみに使用します。
+- 問い合わせ内容は、本サービスの維持改善に役立てる目的のみに使用します。
+
+#### FCMトークンについて
+本サービスは、FCMトークンを収集します。  
+本サービスの通知を許可することで、本サービスのサーバーにFCMトークンが送信されます。
+FCMトークンには、ユーザーの氏名やメールアドレスなど、個人を特定できる情報は含まれておらず、FCMトークンを使ってユーザーのデバイスに通知を送ることしかできません。
+また、FCMトークンはユーザー側で通知のブロック、通知の解除をすることでいつでも無効化することができます。
+
+#### JavaScriptの利用について
+本サービスでは、通知を受け取るようにするため、JavaScriptを使用しております。  
+JavaScriptを無効にすると、本サービスをご利用できませんので、JavaScriptを有効にしていただけますようお願いいたします。
+
+#### 問い合わせについて
+Googleフォームから問い合わせの際、問い合わせ内容が本サービスのサーバーに送信されます。
+問い合わせの際、Googleアカウントが必要ですが、ユーザーのGoogleアカウント情報は収集いたしません。
+
+#### 個人情報の第三者提供
+ユーザーからお預かりした個人情報を、個人情報保護法その他の法令に基づき開示が認められる場合を除き、ご本人様の同意を得ずに第三者に提供することはありません。
+
+#### 本サービスのコンテンツ
+本サービスは、YouTube APIサービスによって動画情報の取得・通知をおこなっています。
+また、YouTube APIサービスの利用は、YouTubeの[利用規約](https://www.youtube.com/t/terms)及びGoogleの「[プライバシーポリシー](https://policies.google.com/privacy)」に則り利用しております。  
+ユーザーは本サービスを利用することでYouTubeの[利用規約](https://www.youtube.com/t/terms)及びGoogleの「[プライバシーポリシー](https://policies.google.com/privacy)に同意したものとみなされます。
+
+#### プライバシーポリシーの変更
+必要に応じて、このプライバシーポリシーの内容を変更します。この場合、変更後のプライバシーポリシーの施行時期と内容を適切な方法により周知または通知します。
+
+
+2024年05月12日 制定
