+++ frontend/tsconfig.json
@@ -1,0 +1,8 @@
+{
+"extends": "astro/tsconfigs/strict",
+"compilerOptions": {
+    "types": [
+        "@cloudflare/workers-types/2023-07-01"
+    ]
+}
+}
