--- frontend/vite.config.js
@@ -1,16 +1,0 @@
-import { defineConfig } from 'vite';
-
-export default defineConfig({
-  base: '/',
-  build: {
-    rollupOptions: {
-      input: [
-        'index.html',
-        'firebase-messaging-sw.js'
-      ],
-      output: {
-        entryFileNames: '[name].js',
-      }
-    }
-  }
-});
