+++ frontend/wrangler.toml
@@ -2,9 +2,18 @@ name = "niji-tuu"
 compatibility_date = "2024-04-05"
 pages_build_output_dir = "./dist"
 
+# Automatically place your workloads in an optimal location to minimize latency.
+# If you are running back-end logic in a Pages Function, running it closer to your back-end infrastructure
+# rather than the end user may result in better performance.
+# Docs: https://developers.cloudflare.com/pages/functions/smart-placement/#smart-placement
+# [placement]
+# mode = "smart"
+
 # Variable bindings. These are arbitrary, plaintext strings (similar to environment variables)
+# Docs:
+# - https://developers.cloudflare.com/pages/functions/bindings/#environment-variables
 # Note: Use secrets to store sensitive data.
-# Docs: https://developers.cloudflare.com/pages/functions/bindings/#environment-variables
+# - https://developers.cloudflare.com/pages/functions/bindings/#secrets
 # [vars]
 # MY_VARIABLE = "production_value"
 
@@ -17,8 +26,12 @@ pages_build_output_dir = "./dist"
 # Docs: https://developers.cloudflare.com/pages/functions/bindings/#d1-databases
 [[d1_databases]]
 binding = "MY_DB"
-database_name = "niji-tuu"
-database_id = "f7b41d86-2565-4fca-83fb-2bab1bb291c8"
+# database_name = "niji-tuu"
+# database_id = "f7b41d86-2565-4fca-83fb-2bab1bb291c8"
+
+database_name = "niji-tuu-dev"
+database_id = "20be5d67-9413-49be-ae55-64511ffe8231"
+
 preview_database_id = "DB"
 
 # Bind a Durable Object. Durable objects are a scale-to-zero compute primitive based on the actor model.
@@ -32,8 +45,8 @@ preview_database_id = "DB"
 # Bind a KV Namespace. Use KV as persistent storage for small key-value pairs.
 # Docs: https://developers.cloudflare.com/pages/functions/bindings/#kv-namespaces
 # [[kv_namespaces]]
-# binding = "first_kv"
-# id = "4edba682067e463a9b6762b81d435631"
+# binding = "MY_KV_NAMESPACE"
+# id = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
 
 # Bind a Queue producer. Use this binding to schedule an arbitrary task that may be processed later by a Queue consumer.
 # Docs: https://developers.cloudflare.com/pages/functions/bindings/#queue-producers
@@ -52,3 +65,25 @@ preview_database_id = "DB"
 # [[services]]
 # binding = "MY_SERVICE"
 # service = "my-service"
+
+# To use different bindings for preview and production environments, follow the examples below.
+# When using environment-specific overrides for bindings, ALL bindings must be specified on a per-environment basis.
+# Docs: https://developers.cloudflare.com/pages/functions/wrangler-configuration#environment-specific-overrides
+
+######## PREVIEW environment config ########
+
+# [env.preview.vars]
+# API_KEY = "xyz789"
+
+# [[env.preview.kv_namespaces]]
+# binding = "MY_KV_NAMESPACE"
+# id = "<PREVIEW_NAMESPACE_ID>"
+
+######## PRODUCTION environment config ########
+
+# [env.production.vars]
+# API_KEY = "abc123"
+
+# [[env.production.kv_namespaces]]
+# binding = "MY_KV_NAMESPACE"
+# id = "<PRODUCTION_NAMESPACE_ID>"
